pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "anima-lsfdnet"
version = "0.1.0"
description = "ANIMA-ready LSFDNet fusion core implementation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.25",
  "Pillow>=10.0",
  "torch>=2.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
api = [
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "python-multipart>=0.0.9",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/anima_lsfdnet"]
